Refined pathfinding
Pathfinding is now functional. Attach Mover2 and PathfindHelper to an
object (Making sure to set the p variable in mover to the
PathfindHelper), set a destination, then set "found" to false (I'm not
uploading a scene because I don't want to have that whole "detached
head" thing happen again, since I don't know how to fix it).

You can demo it by attaching it to an object and then manually
unchecking the box in the editor.

Some odd behaviors if the start point or destination is occupied, but
there's no reason for either of those cases to occur.

Note that the expected size of a character is 0.5x0.5, larger characters
won't fit within a single "square".

Additionally, I made a few miscellaneous changes to MapGeneration3,
which should hopefully make it a bit easier to work with.